fixed amount rounding issue, tests and transfer amount bug

+++ banking-account-api/src/main/kotlin/com/jpan/banking_account_api/dto/AccountBalanceDto.kt
@@ -1,9 +1,10 @@
 package com.jpan.banking_account_api.dto
 
 import com.jpan.banking_account_api.model.AccountId
+import java.math.BigDecimal
 
 data class AccountBalanceDto(
     val account: AccountId,
     val userLastName: String,
-    val balance: Double
+    val balance: BigDecimal
 )
